the-dynamodb-atomic-counter/python/requirements.txt
-e .
aws-cdk.assets==1.83.0
aws-cdk.aws-iam==1.83.0
aws-cdk.aws-apigateway==1.83.0
aws-cdk.aws-dynamodb==1.83.0
aws-cdk.core==1.83.0
aws-cdk.custom_resources==1.83.0